Déplacement des composants de la page Home, Correctif des import et refactorisation
