+++ client/src/components/Public/HomeRoot/AvisClientHomeControllerData.tsx
@@ -1,4 +1,4 @@
-import AvisClientControllerDataType from "../../types/AvisClientControllerData.type";
+import AvisClientControllerDataType from "../../../types/AvisClientControllerData.type";
 
 const AvisClientHomeControllerData: AvisClientControllerDataType[] = [
     /* Commentaire 1 */
